+++ djpcms/views/apps/tagging/__init__.py
@@ -107,11 +107,11 @@ class ArchiveTaggedApplication(ArchiveApplication):
     month_archive1 = TagArchiveView(regex = '(?P<month>\w{3})', parent = 'year_archive1')
     day_archive1   = TagArchiveView(regex = '(?P<day>\d{2})',   parent = 'month_archive1')
     
-    tagc1          = appview.AppView(regex = 'tags2/(?P<tag1>%s)' % tag_regex)
-    tag2           = TagArchiveView(regex = '(?P<tag2>%s)' % tag_regex, parent = 'tagc1')
-    year_archive2  = TagArchiveView(regex = '(?P<year>\d{4})',  parent = 'tag2')
-    month_archive2 = TagArchiveView(regex = '(?P<month>\w{3})', parent = 'year_archive2')
-    day_archive2   = TagArchiveView(regex = '(?P<day>\d{2})',   parent = 'month_archive2')
+    #tagc1          = appview.AppView(regex = 'tags2/(?P<tag1>%s)' % tag_regex)
+    #tag2           = TagArchiveView(regex = '(?P<tag2>%s)' % tag_regex, parent = 'tagc1')
+    #year_archive2  = TagArchiveView(regex = '(?P<year>\d{4})',  parent = 'tag2')
+    #month_archive2 = TagArchiveView(regex = '(?P<month>\w{3})', parent = 'year_archive2')
+    #day_archive2   = TagArchiveView(regex = '(?P<day>\d{2})',   parent = 'month_archive2')
     
     def tagurl(self, request, *tags):
         return tagurl(self, request, *tags)